feat: Add support for multiple state cycles
This commit introduces the ability to have multiple cycles of checkbox
states. The `states` configuration option can now be a list of lists,
with each sublist representing a separate cycle of states. The
`find_cycle_index` function has been added to determine the current
cycle and state index. The `cycle_state` function has been updated to
use this new structure.